+++ src/common/services/CriteriaService.ts
@@ -1,7 +1,7 @@
 import { DevOpsService } from '@joachimdalen/azdevops-ext-core/DevOpsService';
 import { getLoggedInUser } from '@joachimdalen/azdevops-ext-core/IdentityUtils';
 import { getClient } from 'azure-devops-extension-api';
-import { CoreRestClient, Process, WebApiTeam } from 'azure-devops-extension-api/Core';
+import { CoreRestClient, WebApiTeam } from 'azure-devops-extension-api/Core';
 import { GraphMembership, GraphRestClient } from 'azure-devops-extension-api/Graph';
 import {
   WorkItemQueryResult,
@@ -16,7 +16,6 @@ import {
   CriteriaDocument,
   CriteriaPanelConfig,
   FullCriteriaStatus,
-  HistoryEvent,
   HistoryItem,
   IAcceptanceCriteria,
   ProcessEvent
@@ -197,6 +196,7 @@ class CriteriaService {
           criteria.state = AcceptanceCriteriaState.New;
         }
         details.processed = undefined;
+        details.latestComment = undefined;
       } else {
         if (criteria.requiredApprover) {
           criteria.state = AcceptanceCriteriaState.AwaitingApproval;
@@ -216,7 +216,8 @@ class CriteriaService {
   public async processCriteria(
     workItemId: string,
     id: string,
-    action: ProcessEvent
+    action: ProcessEvent,
+    comment?: string
   ): Promise<{ criteria: IAcceptanceCriteria; details: CriteriaDetailDocument } | undefined> {
     try {
       const doc = await this._dataStore.getCriteriasForWorkItem(workItemId);
@@ -238,7 +239,7 @@ class CriteriaService {
               processedBy: approver
             };
           }
-
+          details.latestComment = comment;
           criteria.state =
             action === ProcessEvent.Approve
               ? AcceptanceCriteriaState.Approved
@@ -246,7 +247,11 @@ class CriteriaService {
 
           const res = await this.update(doc, criteria, details);
 
-          const historyEvent: HistoryItem = this._historyService.getProcessEvent(action, approver);
+          const historyEvent: HistoryItem = this._historyService.getProcessEvent(
+            action,
+            approver,
+            comment
+          );
           await this._historyService.createOrUpdate(id, historyEvent);
           this.emitChange(false, true);
           return res;